requirements.txt
pandas
bs4
tweepy
TurkishStemmer
grpcio
price_parser
zemberek_grpc
multi_rake
sklearn
pytorch-transformers==1.2.0
transformers==2.9.0